+++ src/config/constants.ts
@@ -4,17 +4,8 @@ const ASYNC_SCRIPTS = 'async';
 
 const STATIC_ASSETS = 'assets';
 
-const MANIFEST = 'manifest.json';
-
 const INJECT_HASHNAME = '[contenthash:8]';
 
 const HMR_PATH = '/__balm_hmr';
 
-export {
-  ASSETS_KEYS,
-  ASYNC_SCRIPTS,
-  STATIC_ASSETS,
-  MANIFEST,
-  INJECT_HASHNAME,
-  HMR_PATH
-};
+export { ASSETS_KEYS, ASYNC_SCRIPTS, STATIC_ASSETS, INJECT_HASHNAME, HMR_PATH };