+++ src/config/extras.ts
@@ -1,7 +1,5 @@
-const includes: string[] = [];
 const excludes: string[] = [];
 
 export default {
-  includes,
   excludes
 };